Rename package highlightextraction to frequency.
