+++ core/src/main/java/de/jplag/JPlag.java
@@ -15,7 +15,7 @@ import de.jplag.comparison.LongestCommonSubsequenceSearch;
 import de.jplag.exceptions.ExitException;
 import de.jplag.exceptions.RootDirectoryException;
 import de.jplag.exceptions.SubmissionException;
-import de.jplag.highlightextraction.MatchWeighting;
+import de.jplag.frequency.MatchWeighting;
 import de.jplag.merging.MatchMerging;
 import de.jplag.options.JPlagOptions;
