+++ core/src/main/java/de/jplag/frequency/SublistIterator.java
@@ -1,4 +1,4 @@
-package de.jplag.highlightextraction;
+package de.jplag.frequency;
 
 import java.util.Iterator;
 import java.util.List;
