+++ core/src/main/java/de/jplag/frequency/strategy/CompleteMatchesStrategy.java
@@ -1,4 +1,4 @@
-package de.jplag.highlightextraction.strategy;
+package de.jplag.frequency.strategy;
 
 import java.util.List;
