+++ core/src/main/java/de/jplag/frequency/strategy/ContainedMatchesStrategy.java
@@ -1,11 +1,11 @@
-package de.jplag.highlightextraction.strategy;
+package de.jplag.frequency.strategy;
 
 import java.util.Iterator;
 import java.util.List;
 import java.util.stream.IntStream;
 
 import de.jplag.TokenType;
-import de.jplag.highlightextraction.SublistIterator;
+import de.jplag.frequency.SublistIterator;
 
 /**
  * Strategy that counts all occurrences of complete matches inside all complete matches and contiguous submatches from
