+++ core/src/main/java/de/jplag/frequency/weighting/ProportionalWeighting.java
@@ -1,6 +1,6 @@
-package de.jplag.highlightextraction.weighting;
+package de.jplag.frequency.weighting;
 
-import de.jplag.highlightextraction.WeightingFunction;
+import de.jplag.frequency.WeightingFunction;
 
 /**
  * Frequency weighting strategy which emphasizes or de-emphasizes matches proportionally in terms of their rarity.
