+++ core/src/test/java/de/jplag/frequency/WeightStrategyTest.java
@@ -1,4 +1,4 @@
-package de.jplag.highlightextraction;
+package de.jplag.frequency;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -19,11 +19,11 @@ import de.jplag.TestBase;
 import de.jplag.TokenType;
 import de.jplag.comparison.LongestCommonSubsequenceSearch;
 import de.jplag.exceptions.ExitException;
-import de.jplag.highlightextraction.strategy.CompleteMatchesStrategy;
-import de.jplag.highlightextraction.strategy.ContainedMatchesStrategy;
-import de.jplag.highlightextraction.strategy.FrequencyStrategy;
-import de.jplag.highlightextraction.strategy.SubmatchesStrategy;
-import de.jplag.highlightextraction.strategy.WindowOfMatchesStrategy;
+import de.jplag.frequency.strategy.CompleteMatchesStrategy;
+import de.jplag.frequency.strategy.ContainedMatchesStrategy;
+import de.jplag.frequency.strategy.FrequencyStrategy;
+import de.jplag.frequency.strategy.SubmatchesStrategy;
+import de.jplag.frequency.strategy.WindowOfMatchesStrategy;
 import de.jplag.options.JPlagOptions;
 
 /**
