+++ core/src/test/java/de/jplag/frequency/strategy/StrategyTest.java
@@ -1,4 +1,4 @@
-package de.jplag.highlightextraction.strategy;
+package de.jplag.frequency.strategy;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -25,7 +25,7 @@ import de.jplag.Token;
 import de.jplag.TokenType;
 import de.jplag.comparison.LongestCommonSubsequenceSearch;
 import de.jplag.exceptions.ExitException;
-import de.jplag.highlightextraction.TokenSequenceUtil;
+import de.jplag.frequency.TokenSequenceUtil;
 import de.jplag.options.JPlagOptions;
 
 /**
